Part 2 of implementing #393
